+++ GetThePicture.Tests/Copybook/Compiler/Base/ParserForLv88Test.cs
@@ -22,14 +22,15 @@ public class ParserForLv88Test
 
         Parser parser = new(tokens);
 
-        var model = parser.Analyze();
+        var layout = parser.Analyze();
+        layout.Seal();
         
-        Assert.IsNotNull(model);
+        Assert.IsNotNull(layout);
 
         var sb = new StringBuilder();
         using var writer = new StringWriter(sb);
 
-        model.Dump(writer);
+        layout.Dump(writer);
 
         string result = sb.ToString();